fix(profiling): move profile percentage increment from shouldProfile
BREAKING CHANGE: shouldProfile() previously had a side effect where it
was incrementing the profile percentage every time it was called. This
would be problematic, for example, if for some reason the method was called
twice in the same invocation. So it was moved to init() after the
invocation of shouldProfile().