refactor(invocation): clean up invocation process
BREAKING CHANGE: We no longer call cleanup() twice if an originally
successful request fails during the respond() step.

An overly-cautious fallback was added for cases where the respond method
is throwing errors. We try to invoke it with the successful response,
then with the resulting error, then we fallback to invoking callback
directly.