refactor(handler): add tests and other misc fixes/refactors (#9)
* feat(handler): track totalInvocations

* fix(profiling): move profile percentage increment from shouldProfile

BREAKING CHANGE: shouldProfile() previously had a side effect where it
was incrementing the profile percentage every time it was called. This
would be problematic, for example, if for some reason the method was called
twice in the same invocation. So it was moved to init() after the
invocation of shouldProfile().

* refactor(profiling): use total invocations for ONE_COLD_ONE_WARM

BREAKING CHANGE: It is unlikely to have any impact since the end result
is the same, but any implementation previously relying on the total of
the profile percentage incrementor will break.

* refactor(invocation): clean up invocation process

BREAKING CHANGE: We no longer call cleanup() twice if an originally
successful request fails during the respond() step.

An overly-cautious fallback was added for cases where the respond method
is throwing errors. We try to invoke it with the successful response,
then with the resulting error, then we fallback to invoking callback
directly.

* feat(tests): add tests! :100: